details_page_2

+++ app/src/main/java/tech/libin/rahul/ideaproject/views/models/ActivityRequestModel.java
@@ -8,7 +8,7 @@ import tech.libin.rahul.ideaproject.configurations.Constants;
 
 public class ActivityRequestModel {
 
-    private int userId;
+    private Long userId;
 
     private int pageNo;
 
@@ -24,11 +24,11 @@ public class ActivityRequestModel {
 
     private String name;
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
